+++ crates/cubek-matmul/src/eval/benchmarks/gemm_cpu/mod.rs
@@ -5,24 +5,21 @@ use cubek_test_utils::{CatalogEntry, RunSamples};
 use crate::eval::benchmarks::gemm::{self, GemmProblem};
 use crate::strategy::Strategy;
 
-/// CpuGemm vs the unit matmuls, plus the forced-tile mask probe. The forced tiles
-/// are diagnostic on the 512 square: `t64`/`t32` divide 512 (maskless), `t48` does
-/// not (masked).
+/// CpuGemm vs the simple-unit baseline, plus the forced-tile mask probes. The forced
+/// tiles are diagnostic on the 512 square: `t64`/`t32` divide 512 (maskless), `t48`
+/// does not (masked).
 const STRATEGIES: &[&str] = &[
     "cpu_gemm",
     "simple_unit_min",
-    "double_unit_min",
     "cpu_gemm_t48",
     "cpu_gemm_t64",
     "cpu_gemm_t32",
 ];
 
-/// Aligned square, a vector × matrix, and a matrix × vector — CPU-sized shapes.
-const PROBLEMS: &[&str] = &[
-    "rect_1x512x512x512_rr_f32",
-    "vecmat_2x1x4096x4096_rr_f32",
-    "matvec_2x8192x1x8192_rr_f32",
-];
+/// Base shapes; the catalog expands each over all four layouts (rr/rc/cr/cc) and both
+/// precisions (f32/f16). The 512 square keeps the CPU reference cheap while still
+/// exercising the vectorized (row-major) and scalar (col-major) paths under masking.
+const SHAPES: &[&str] = &["rect_1x512x512x512"];
 
 pub struct Category;
 
@@ -41,7 +38,7 @@ impl cubek_test_utils::Category for Category {
     fn problems(&self) -> Vec<CatalogEntry<GemmProblem>> {
         gemm::problems()
             .into_iter()
-            .filter(|p| PROBLEMS.contains(&p.id.as_str()))
+            .filter(|p| SHAPES.iter().any(|s| p.id.starts_with(&format!("{s}_"))))
             .collect()
     }
